Move unit test to foundation namespace
